Modify: Using Singleton pattern into the Controller Manager
