+++ src/TestSmartDeviceAPPs/iOS/IoTDeviceController/IoTDeviceController/Notification.swift
@@ -1,0 +1,13 @@
+//
+//  Notification.swift
+//  IoTDeviceController
+//
+//  Created by Kim, Min Ho on 2016. 2. 18..
+//  Copyright © 2016년 Luna. All rights reserved.
+//
+
+import Foundation
+
+protocol Notification {
+    func doSomething()
+}
